feat(order): implementar verification de precios, descuentos, y resolver bugs relacionados
- Agregar verificaciones de datos de frontend vs backend para prevenir manipulacion de precios.
- Implementar calculo real de descuentos sobre los precios (precio final).
- Resolver bugs relacionados con el procesamiento de ordenes.
- Mover logica de ListingMapper a Helper dedicado
- Mover logica de OrderMapper a Helper dedicado
- Resolver bugs de relacionados con favorites.

+++ src/main/java/com/techlab/store/dto/CreateListingDTO.java
@@ -1,13 +1,15 @@
 package com.techlab.store.dto;
 import com.techlab.store.enums.ListingStatus;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CreateListingDTO (
     // Post 
     String       title,
     String       description,
-    Double       price,
-    Double       discountPercentage,
+    BigDecimal   price,
+    Integer      discountPercentage,
     String       warrantyInformation,
     String       shippingInformation,
     String       returnPolicy,
